requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.30
matplotlib>=3.7.0
scipy>=1.10.0
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0
openpyxl>=3.1.0
lxml>=4.9.0
pytest>=7.0.0
